Struct windows_acl::acl::ACLEntry

pub struct ACLEntry {
    pub index: u16,
    pub entry_type: AceType,
    pub entry_size: DWORD,
    pub size: WORD,
    pub flags: BYTE,
    pub mask: ACCESS_MASK,
    pub sid: Option<Vec<u16>>,
    pub string_sid: String,
}

ACLEntry represents a single access control entry in an access control list

Fields
index: u16

The index of the current entry in the raw access control list

entry_type: AceType

The entry's type

entry_size: DWORD

The calculated size of the current access control entry

size: WORD

See AceSize in ACE_HEADER

flags: BYTE

See AceFlags in ACE_HEADER

mask: ACCESS_MASK

See ACCESS_MASK

sid: Option<Vec<u16>>

The target entity's raw SID

string_sid: String

The target entity's SID in string representation

Methods

impl ACLEntry

pub fn new() -> ACLEntry

Returns an ACLEntry object with default values.